+++ app/Http/Controllers/DirectMessageController.php
@@ -704,7 +704,7 @@ class DirectMessageController extends Controller
 	public function remoteDeliver($dm)
 	{
 		$profile = $dm->author;
-		$url = $dm->recipient->inbox_url;
+		$url = $dm->recipient->sharedInbox ?? $dm->recipient->inbox_url;
 
 		$tags = [
 			[
@@ -760,7 +760,7 @@ class DirectMessageController extends Controller
 	public function remoteDelete($dm)
 	{
 		$profile = $dm->author;
-		$url = $dm->recipient->inbox_url;
+		$url = $dm->recipient->sharedInbox ?? $dm->recipient->inbox_url;
 
 		$body = [
 			'@context' => [