+++ app/Jobs/InboxPipeline/InboxValidator.php
@@ -9,10 +9,10 @@ use App\Util\ActivityPub\{
     Inbox
 };
 use Illuminate\Bus\Queueable;
-use Illuminate\Queue\SerializesModels;
-use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Queue\SerializesModels;
 use Zttp\Zttp;
 
 class InboxValidator implements ShouldQueue