+++ app/Transformer/ActivityPub/ProfileTransformer.php
@@ -44,6 +44,9 @@ class ProfileTransformer extends Fractal\TransformerAbstract
             'mediaType' => 'image/jpeg',
             'url'       => $profile->avatarUrl(),
           ],
+          'endpoints' => [
+            'sharedInbox' => config('app.url') . '/f/inbox'
+          ]
       ];
     }
 }