+++ config/federation.php
@@ -15,12 +15,12 @@ return [
 		'enabled' => env('ACTIVITY_PUB', false),
 		'outbox' => env('AP_OUTBOX', true),
 		'inbox' => env('AP_INBOX', true),
-		'sharedInbox' => env('AP_SHAREDINBOX', false),
+		'sharedInbox' => env('AP_SHAREDINBOX', true),
 
 		'remoteFollow' => env('AP_REMOTE_FOLLOW', false),
 
 		'delivery' => [
-			'timeout' => env('ACTIVITYPUB_DELIVERY_TIMEOUT', 2.0),
+			'timeout' => env('ACTIVITYPUB_DELIVERY_TIMEOUT', 30.0),
 			'concurrency' => env('ACTIVITYPUB_DELIVERY_CONCURRENCY', 10),
 			'logger' => [
 				'enabled' => env('AP_LOGGER_ENABLED', false),